Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
 int main() {
-    float radius;
+    double radius;
     double Area;
     printf("Area: ");
-    scanf("%f",&radius);
+    scanf("%lf",&radius);
     Area=(M_PI*radius*radius);
     printf("%f",Area);
     return 0;
